Correction des fonctions : 	READER_UTILS_ComputeBaudRate() 	READER_UTILS_GetCardFreq() 	READER_UTILS_ComputeNewBaudRate() pour mieux gérer les float.

+++ src/reader_utils.c
@@ -4,15 +4,15 @@
 
 
 uint32_t READER_UTILS_ComputeBaudRate(uint32_t freq, uint32_t Fi, uint32_t Di){
-	return freq / (Fi/Di);
+	return (uint32_t)((float)freq / ((float)Fi/(float)Di));
 }
 
 uint32_t READER_UTILS_GetCardFreq(uint32_t SYSCLK, uint32_t AHB, uint32_t APB1, uint32_t value_USART_GTPR_PSC){
-	return SYSCLK / AHB / APB1 / (value_USART_GTPR_PSC * 2);   /* Voir en.DM00031020 section 30.6.7  et en.DM00031020 section 7.2 figure 21 */
+	return (float)SYSCLK / (float)AHB / (float)APB1 / (float)(value_USART_GTPR_PSC * 2);   /* Voir en.DM00031020 section 30.6.7  et en.DM00031020 section 7.2 figure 21 */
 }
 
 uint32_t READER_UTILS_ComputeNewBaudRate(uint32_t oldBaudRate, uint32_t oldFreq, uint32_t newFreq){
-	return (newFreq / oldFreq) * oldBaudRate;
+	return (uint32_t)(((float)newFreq / (float)oldFreq) * (float)oldBaudRate);
 }
 
 uint32_t READER_UTILS_ComputePrescFromFreq(uint32_t freq){
